+++ docs/CHANGELOG.md
@@ -3,8 +3,10 @@
 ### v0.3.2 (dev)
 - Set up generic compiler to generalize classes
 - Simplified step evaluation process
+- Added expressions: `Set`
 - Added functions:
   - Numeric: `Factor`
+  - Structural: `SetBracket`
 - Shrunk package size
 ### v0.3.1 (Nov 22, 2021)
 - Set up simple user interface in web